Drinking app

+++ Drinking/app/src/main/java/com/example/jarek/drinking/Math.java
@@ -4,7 +4,7 @@ package com.example.jarek.drinking;
  * Created by Jarek on 1/19/2015.
  */
 public class Math {
-    public double setAlcohol(int A, int W, double r,double mr){
+    public double setAlcohol(double A, int W, double r,double mr){
         double H = 1;
         double bac = ((A * 5.14) /(W * r))-(mr * H);
         return bac;
